requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib.3.8.0
seaborn==0.12.2
streamlit==1.27.1
japanize-matplotlib==1.1.3
requests==2.31.0